// Samples/TS/Angular2/MultiRowIntro/MultiRowIntro/src/services/DataSvc.d.ts
export declare class DataSvc {
    cityMap: wijmo.grid.DataMap;
    orders: any[];
    groupedOrders: wijmo.collections.CollectionView;
    pagedOrders: wijmo.collections.CollectionView;
    addNewOrders: wijmo.collections.CollectionView;
    constructor();
    randBetween(min: any, max: any): number;
}